TS-39576: Move extract_build_errors to dedicated utility file
